requirements.txt
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
bcrypt==3.1.7
boto3==1.10.22
botocore==1.13.22
celery==4.4.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coverage==4.5.4
cx-Oracle==7.2.3
decorator==4.4.0
django-cacheops==4.2
django-cors-headers==3.1.0
django-cors-middleware==1.4.0
django-db-connection-pool==1.0.2
django-debug-toolbar==2.0
django-enumfield==1.5
django-oauth-toolkit==1.2.0
django-pickling==1.0
django-profile-middleware==0.1
django-redis==4.10.0
django-silk==3.0.2
djangorestframework==3.10.2
docutils==0.15.2
fpdf==1.7.2
funcy==1.13
futures==3.1.1
fuzzysearch==0.6.2
fuzzywuzzy==0.17.0
gprof2dot==2017.9.19
gensim==3.8.1
graphene==2.1.5
graphene-django==2.3.0
graphql-core==2.2
graphql-relay==0.4.5
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
JSON-log-formatter==0.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.1.1
newrelic==5.2.1.129
nltk==3.4.5
numpy==1.17.0
oauthlib==3.1.0
pandas==0.23.0
parso==0.5.1
pdfkit==0.6.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pycryptodome==3.9.4
pycparser==2.19
Pygments==2.4.2
pyenchant==3.2.2
pylint==2.3.1
PyMySQL==0.9.3
python-consul==1.1.0
python-dateutil==2.8.0
python-oauth2==1.1.1
pytz==2019.2
redis==3.3.8
regex==2019.6.8
requests==2.22.0
s3transfer==0.2.1
sentry-sdk==0.20.3
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
spacy==2.2.4
traitlets==4.3.2
typed-ast==1.4.0
textblob==0.15.3
urllib3==1.25.3
uWSGI==2.0.18
wordcloud==1.6.0
wcwidth==0.1.7
wrapt==1.11.2
yapf==0.28.0
python3-saml==1.9.0
onelogin==1.9.0
zipp==0.6.0
requests-aws4auth==0.9
elasticsearch==7.5.1
django-prometheus==2.0.0
pre-commit==2.6.0
firebase-admin==4.3.0
PyJWT==1.7.1
python-Levenshtein==0.12.0
cryptography==3.0
pysftp==0.2.8
django-jsonfield-backport==1.0.4
python-gnupg==0.4.8
python-pptx==0.6.21
enum34==1.1.10
phonenumbers==8.12.54
pycurl==7.45.2
